requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-hot-reload==0.3.0
flask-sock==0.7.0
h11==0.16.0
hjson==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
simple-websocket==1.1.0
soupsieve==2.8
typing_extensions==4.15.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
beautifulsoup4==4.14.3
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-hot-reload==0.3.0
flask-sock==0.7.0
h11==0.16.0
hjson==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
simple-websocket==1.1.0
soupsieve==2.8
typing_extensions==4.15.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
